+++ src/components/Header/elements/Blank.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+const Blank = styled.div`
+    position: fixed;
+    z-index: 1;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 100%;
+    opacity: 0;
+`;
+
+export default Blank;
